+++ src/lib/CoinbasePro/Authenticated/Deposit.hs
@@ -3,16 +3,13 @@
 {-# LANGUAGE TemplateHaskell            #-}
 
 module CoinbasePro.Authenticated.Deposit
-    ( Deposit (..)
+    ( DepositDetails (..)
     , DepositRequest (..)
     , DepositResponse (..)
     , CoinbaseDepositRequest (..)
     , CryptoDepositAddress (..)
     ) where
 
-import           Data.Aeson                         (FromJSON, ToJSON,
-                                                     parseJSON, withObject,
-                                                     (.:), (.:?))
 import           Data.Aeson.Casing                  (snakeCase)
 import           Data.Aeson.TH                      (defaultOptions, deriveJSON,
                                                      fieldLabelModifier)
@@ -22,54 +19,20 @@ import           Data.UUID                          (UUID)
 
 import           CoinbasePro.Authenticated.Accounts (AccountId)
 import           CoinbasePro.Authenticated.Payment  (PaymentMethodId (..))
-import           CoinbasePro.Types                  (CreatedAt, UserId)
+import           CoinbasePro.Types                  (CreatedAt)
 
 
 data DepositDetails = DepositDetails
-    { cryptoAddress         :: Text
-    , destinationTag        :: Text
+    { cryptoAddress         :: Maybe Text
+    , destinationTag        :: Maybe Text
     , coinbaseAccountId     :: AccountId
-    , destinationTagName    :: Text
+    , destinationTagName    :: Maybe Text
     , coinbaseTransactionId :: Text
-    , cryptoTransactionHash :: Text
+    , cryptoTransactionHash :: Maybe Text
     } deriving (Eq, Show)
 
 
 deriveJSON defaultOptions { fieldLabelModifier = snakeCase } ''DepositDetails
-
-
-newtype UserNonce = UserNonce Int
-  deriving (Eq, Show, ToJSON, FromJSON)
-
-
-data Deposit = Deposit
-    { dId          :: PaymentMethodId
-    , dType        :: Text
-    , dCreatedAt   :: CreatedAt
-    , dCompletedAt :: UTCTime
-    , dCanceledAt  :: Maybe UTCTime
-    , dProcessedAt :: Maybe UTCTime
-    , dAccountId   :: AccountId
-    , dUserId      :: UserId
-    , dUserNonce   :: Maybe UserNonce
-    , dAmount      :: Double
-    , dDetails     :: Maybe DepositDetails
-    } deriving (Eq, Show)
-
-
-instance FromJSON Deposit where
-  parseJSON = withObject "deposit" $ \o -> Deposit
-    <$> (o .: "id")
-    <*> (o .: "type")
-    <*> (o .: "created_at")
-    <*> (o .: "completed_at")
-    <*> (o .:? "canceled_at")
-    <*> (o .:? "processed_at")
-    <*> (o .: "account_id")
-    <*> (o .: "user_id")
-    <*> (o .:? "user_nonce")
-    <*> (read <$> o .: "amount")
-    <*> (o .:? "deposit_details")
 
 
 data DepositRequest = DepositRequest